fix: search the point of failure in the 'all' scope
The test might fail to initialize in a class from a library as it was
the case for #319. Search in the 'all' scope (including libraries) for
the point of failure. Since we might land in a decompiled class the line
number might be outside of the range for the decompiled source, so we
need to adjust it or we get an exception in the highlighter pass